% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emisje.R
\docType{package}
\name{Emisja}
\alias{Emisja}
\title{Pakiet Emisje}
\description{
Pakiet oblicza emisje EMEP/EEA zanieczyszczenia powietrza wybranego rodzaju pojazdu,
a takze wizualizujacy za pomoca wykresu wartosci wynikowe.
}
\details{
Pakiet zawiera dwie funkcje:

\code{fun_obl_em()} oblicza emisje

\code{fun_em_plot()} rysuje wykresy emisji
}
\section{Dane do obliczen}{

Funkcje pakietu korzystaja z danych dolaczonych do pakietu.

Dane sa dostepne dla uzytkownika jako obiekty \code{input}, \code{wskazniki}.
}

