% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_obl_em.R
\name{fun_obl_em}
\alias{fun_obl_em}
\title{Title}
\usage{
fun_obl_em(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("CO", "EC")
)
}
\arguments{
\item{substancja}{}
}
\value{

}
\description{
Title
}
