% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input}
\alias{input}
\title{Plik zawierajacy dane zewnetrznye koniecznye do dzialania pakietu}
\format{
Opisy znajduja sie w dokumentacji
}
\source{
Obiekt wskazniki zawiera dane pochodzace
z EEA - Europejskiej Agnecji Srodowiska.
Obiekt input zawiera losowe wartosci natezenia ruchu.
}
\usage{
input
}
\description{
Plik zawierajacy dane zewnetrznye koniecznye do dzialania pakietu
}
\examples{
\dontrun{
 input
}
}
\keyword{datasets}
