% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_em_plot.R
\name{fun_em_plot}
\alias{fun_em_plot}
\title{Funkcja wizualizacji wynikow Emisji}
\usage{
fun_em_plot(
  data = wynik,
  kategoria1 = Category,
  wartosci1 = "Passenger Cars",
  kategoria2 = Segment,
  wartosci2 = Emisja,
  fname = "plik.png",
  width = 8,
  height = 6,
  dpi = 300
)
}
\arguments{
\item{data}{data_frame}

\item{kategoria1}{character - wybor pierwszej kolumny z data}

\item{wartosci1}{character - wybor wartosci filtrujacej pierwsza kolumne}

\item{kategoria2}{character - wybor drugiej kolumny}

\item{wartosci2}{double - wybor kolumny z wartosciami double}

\item{fname}{character - nazwa pliku wykresu ktory zostanie zapisany}

\item{width}{double - szerokosc}

\item{height}{double - wysokosc}

\item{dpi}{double - liczba punktow na cal}
}
\value{
plot
}
\description{
Funkcja rysuje wykresy emisji na podstawie danych utworzonych za pomoca \code{fun_obl_em()}
}
\examples{
\dontrun{
# Narysowanie wykresu domyslnego:
fun_em_plot()
# Narysowanie wykresow dla przykladowych danych:
fun_em_plot(dane = wynik,
           kategoria1 = Euro.Standard,
           wartosci1 = "Euro 5",
           kategoria2 = Segment,
           wartosci2 = Nat,
           fname = "plik2.png",
           width = 10,
           height = 8,
           dpi = 600)
}
}
