% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_em_plot.R
\name{fun_em_plot}
\alias{fun_em_plot}
\title{Title}
\usage{
fun_em_plot(
  data = wynik,
  kategoria1 = Category,
  wartosci1 = "Passenger Cars",
  kategoria2 = Segment,
  wartosci2 = Emisja,
  fname = "plik.png",
  width = 8,
  height = 6,
  dpi = 300
)
}
\arguments{
\item{dpi}{}
}
\value{

}
\description{
Title
}
